Changed the logic of the screen buffer
Previously TerminalChar had a prefix and a suffix part which were used
to set and unset the white color of the glowing character. Non-glowing
characters didn't use those fields which mean that most of the memory
consisted of unused zeros.
Now all characters use the prefix to set their own color and suffix was
eliminated. This lead to a ~45% reduction in the size of the screen buffer.
The potential downside is that now an escape sequence is sent before
every character, the performance cost of which is unknown as of yet.

Also aligned all Matrix characters in ALL_MCHARS to the size of the largest
Matrix character. Shorter characters are zero-padded which enables
TerminalChar to copy the mchar without first clearing the MChar part.